+++ tcod/cffi.h
@@ -11,6 +11,7 @@
 #include "../libtcod/src/libtcod/libtcod.h"
 #include "../libtcod/src/libtcod/libtcod_int.h"
 #include "../libtcod/src/libtcod/wrappers.h"
+#include "../libtcod/src/libtcod/tileset/truetype.h"
 
 #include "noise.h"
 #include "path.h"